fix(funding): lock the plan row while changing calendar allocations
The allocation transactions added earlier locked nothing. Allocation
totals are derived by summing sibling rows, so two concurrent changes to
the same plan conflicted on nothing — not the plan row, not the partial
unique index — and each computed its provider total from a snapshot
missing the other's write. Adds ended with more funded calendars than
the provider was billing for; concurrent removals from a two-calendar
plan each saw one allocation remaining, so neither cancelled the plan
and it stayed active funding nothing. Both paths now resolve the plan
FOR UPDATE as the first statement inside their transaction, which
serialises them and lets the second caller's statement-level snapshots
see the first's committed rows. SQLite ignores the lock clause, so this
is a Postgres guarantee.

Corrects the docstrings to claim only what holds. The transaction buys
local atomicity on provider *rejection*; it cannot unwind a provider
mutation, so a commit failure after the provider accepted leaves the
provider and the local rows disagreeing with no compensation path. Also
records why these paths deliberately call the provider inside the
transaction where the rest of the codebase calls out after commit:
nothing else ever recomputes the provider amount from the allocation
rows.

Moves the real-DB rollback tests out of the otherwise fully-stubbed
service unit tests into the domain's integration directory, and adds
direct tests of the deferred-emit helper. The previous "no event on
rollback" assertion was tautological — every injected failure landed
before the emit line — so it passed with a plain synchronous emit. The
new tests fail under that mutation at both the helper and its call site.
