+++ src/server/funding/test/service.test.ts
@@ -9,8 +9,6 @@ import { FundingPlanEntity } from '@/server/funding/entity/funding_plan';
 import { FundingEventEntity } from '@/server/funding/entity/funding_event';
 import { ComplimentaryGrantEntity } from '@/server/funding/entity/complimentary_grant';
 import { CalendarFundingPlanEntity } from '@/server/funding/entity/calendar_funding_plan';
-import { AccountEntity } from '@/server/common/entity/account';
-import { CalendarEntity } from '@/server/calendar/entity/calendar';
 import { ProviderFactory } from '@/server/funding/service/provider/factory';
 import { FundingSettings, ProviderConfig, FundingPlan } from '@/common/model/funding-plan';
 import { ComplimentaryGrant } from '@/common/model/complimentary_grant';
@@ -1465,236 +1463,5 @@ describe('FundingService', () => {
       // processCheckoutCompleted should NOT have been called
       expect(findOneSpy.called).toBe(false);
     });
-  });
-});
-
-/**
- * Transaction boundaries around the allocation paths.
- *
- * These run against real rows (no entity stubs on the writes under test) so a
- * rollback is observable as the absence or unchanged state of a database row.
- * Each test injects a failure at the last step of one wrapped path — always a
- * provider call or a write that happens after at least one local write has
- * already been issued — and asserts nothing partial survived.
- */
-describe('FundingService - allocation transaction rollback', () => {
-  let sandbox: sinon.SinonSandbox;
-  let eventBus: EventEmitter;
-  let service: FundingService;
-  let mockCalendarInterface: {
-    isCalendarOwnerById: sinon.SinonStub;
-    calendarExists: sinon.SinonStub;
-    getCalendarOwnerAccountId: sinon.SinonStub;
-    getCalendar: sinon.SinonStub;
-  };
-
-  let accountId: string;
-  let calendarId: string;
-  let secondCalendarId: string;
-  let providerConfigId: string;
-  let fundingPlanId: string;
-  const periodEnd = new Date(Date.now() + 30 * 24 * 60 * 60 * 1000);
-
-  beforeEach(async () => {
-    await db.sync({ force: true });
-
-    sandbox = sinon.createSandbox();
-    eventBus = new EventEmitter();
-    service = new FundingService(eventBus);
-
-    mockCalendarInterface = {
-      isCalendarOwnerById: sandbox.stub().resolves(true),
-      calendarExists: sandbox.stub().resolves(true),
-      getCalendarOwnerAccountId: sandbox.stub(),
-      getCalendar: sandbox.stub().resolves(null),
-    };
-    service.setCalendarInterface(mockCalendarInterface as any);
-
-    const account = AccountEntity.build({
-      id: uuidv4(),
-      email: 'rollback@example.com',
-      password_hash: 'hash',
-      status: 'active',
-      languages: 'en',
-    });
-    await account.save();
-    accountId = account.id;
-
-    const calendar = CalendarEntity.build({
-      id: uuidv4(),
-      url_name: 'rollback-cal',
-      languages: 'en',
-      default_date_range: 'month',
-    });
-    await calendar.save();
-    calendarId = calendar.id;
-
-    const secondCalendar = CalendarEntity.build({
-      id: uuidv4(),
-      url_name: 'rollback-cal-two',
-      languages: 'en',
-      default_date_range: 'month',
-    });
-    await secondCalendar.save();
-    secondCalendarId = secondCalendar.id;
-
-    const providerModel = new ProviderConfig(uuidv4(), 'stripe');
-    providerModel.enabled = true;
-    providerModel.displayName = 'Credit Card';
-    providerModel.credentials = JSON.stringify({ apiKey: 'sk_test_123' });
-    providerModel.webhookSecret = 'whsec_test_secret';
-    const providerConfig = ProviderConfigEntity.fromModel(providerModel);
-    await providerConfig.save();
-    providerConfigId = providerConfig.id;
-
-    const plan = FundingPlanEntity.build({
-      id: uuidv4(),
-      account_id: accountId,
-      provider_config_id: providerConfigId,
-      provider_subscription_id: 'sub_rollback_existing',
-      provider_customer_id: 'cus_rollback_existing',
-      status: 'active',
-      billing_cycle: 'monthly',
-      amount: 1000000,
-      currency: 'USD',
-      current_period_start: new Date(),
-      current_period_end: periodEnd,
-      cancelled_at: null,
-      suspended_at: null,
-    });
-    await plan.save();
-    fundingPlanId = plan.id;
-  });
-
-  afterEach(() => {
-    sandbox.restore();
-    ProviderFactory.clearAllCaches();
-  });
-
-  async function createAllocation(forCalendarId: string, amount: number): Promise<CalendarFundingPlanEntity> {
-    const allocation = CalendarFundingPlanEntity.build({
-      id: uuidv4(),
-      funding_plan_id: fundingPlanId,
-      calendar_id: forCalendarId,
-      amount,
-      end_time: null,
-    });
-    await allocation.save();
-    return allocation;
-  }
-
-  it('should roll back the new allocation row when the provider amount update fails on add', async () => {
-    const updateSubscriptionAmount = sandbox.stub().rejects(new Error('provider rejected amount update'));
-    sandbox.stub(ProviderFactory, 'getAdapter').returns({
-      supportsAmountUpdates: () => true,
-      updateSubscriptionAmount,
-    } as any);
-
-    await expect(
-      service.addCalendarToFundingPlan(accountId, calendarId, 500000),
-    ).rejects.toThrow('provider rejected amount update');
-
-    // The failure was injected after the allocation row was written
-    expect(updateSubscriptionAmount.calledOnce).toBe(true);
-
-    const allocations = await CalendarFundingPlanEntity.findAll({
-      where: { funding_plan_id: fundingPlanId },
-    });
-    expect(allocations).toHaveLength(0);
-  });
-
-  it('should roll back the end_time write when the provider amount update fails on remove', async () => {
-    const allocation = await createAllocation(calendarId, 500000);
-    await createAllocation(secondCalendarId, 300000);
-
-    const updateSubscriptionAmount = sandbox.stub().rejects(new Error('provider rejected amount update'));
-    sandbox.stub(ProviderFactory, 'getAdapter').returns({
-      supportsAmountUpdates: () => true,
-      updateSubscriptionAmount,
-    } as any);
-
-    await expect(
-      service.removeCalendarFromFundingPlan(accountId, calendarId),
-    ).rejects.toThrow('provider rejected amount update');
-
-    expect(updateSubscriptionAmount.calledOnce).toBe(true);
-
-    await allocation.reload();
-    expect(allocation.end_time).toBeNull();
-  });
-
-  it('should roll back both the end_time write and the plan cancellation when the provider cancel fails', async () => {
-    const allocation = await createAllocation(calendarId, 500000);
-
-    const cancelSubscription = sandbox.stub().rejects(new Error('provider rejected cancellation'));
-    sandbox.stub(ProviderFactory, 'getAdapter').returns({
-      supportsAmountUpdates: () => true,
-      updateSubscriptionAmount: sandbox.stub().resolves(),
-      cancelSubscription,
-    } as any);
-
-    const cancelledEvents: unknown[] = [];
-    eventBus.on('funding:plan:cancelled', (payload) => cancelledEvents.push(payload));
-
-    await expect(
-      service.removeCalendarFromFundingPlan(accountId, calendarId),
-    ).rejects.toThrow('provider rejected cancellation');
-
-    // Last active calendar, so the removal took the cancel branch
-    expect(cancelSubscription.calledOnce).toBe(true);
-
-    await allocation.reload();
-    expect(allocation.end_time).toBeNull();
-
-    const plan = await FundingPlanEntity.findByPk(fundingPlanId);
-    expect(plan!.status).toBe('active');
-    expect(plan!.cancelled_at).toBeNull();
-    expect(cancelledEvents).toHaveLength(0);
-  });
-
-  it('should roll back the funding plan and event rows when a calendar allocation write fails on checkout completion', async () => {
-    sandbox.stub(ProviderFactory, 'getAdapter').returns({
-      getSubscription: sandbox.stub().resolves({
-        providerSubscriptionId: 'sub_rollback_new',
-        providerCustomerId: 'cus_rollback_new',
-        status: 'active',
-        currentPeriodStart: new Date(),
-        currentPeriodEnd: periodEnd,
-        amount: 1000000,
-        currency: 'USD',
-      }),
-    } as any);
-
-    // Fail the last write in the transaction, after the plan and event rows
-    const createStub = sandbox.stub(CalendarFundingPlanEntity, 'create')
-      .rejects(new Error('allocation write failed'));
-
-    const createdEvents: unknown[] = [];
-    eventBus.on('funding:plan:created', (payload) => createdEvents.push(payload));
-
-    await expect(
-      service.processWebhookEvent({
-        eventId: 'evt_rollback_checkout',
-        eventType: 'checkout.session.completed',
-        subscriptionId: 'sub_rollback_new',
-        customerId: 'cus_rollback_new',
-        accountId,
-        calendarIds: JSON.stringify([calendarId]),
-        rawPayload: { type: 'checkout.session.completed' },
-      }, providerConfigId),
-    ).rejects.toThrow('allocation write failed');
-
-    expect(createStub.calledOnce).toBe(true);
-
-    const plans = await FundingPlanEntity.findAll({
-      where: { provider_subscription_id: 'sub_rollback_new' },
-    });
-    expect(plans).toHaveLength(0);
-
-    const events = await FundingEventEntity.findAll({
-      where: { provider_event_id: 'evt_rollback_checkout' },
-    });
-    expect(events).toHaveLength(0);
-    expect(createdEvents).toHaveLength(0);
   });
 });
